[SLP][modularisation][NFC] Extract same-opcode compatibility helpers
Move the SLP same-opcode compatibility primitives out of
SLPVectorizer.cpp into a new private module
SLPVectorizer/SLPCompatibilityAnalysis.{h,cpp}:

  isValidForAlternation
  BinOpSameOpcodeHelper

These do not depend on BoUpSLP. This is a pure code move with no
functional change; the only edit is that the member call to the free
isValidForAlternation is now qualified with slpvectorizer:: (it was
previously ::), since both now live in namespace llvm::slpvectorizer.

Part of the effort to modularize SLPVectorizer.cpp. See the RFC:
https://discourse.llvm.org/t/modularizing-slpvectorizer-cpp/90922
